fix(api): repair the rating filter and page filtered browse queries
- model the public minimum rating as a whole number: IGDB rejects a decimal
  literal in a where clause, so every rating filter failed as
  UPSTREAM_INVALID_RESPONSE, and the adapter test asserted the broken clause
  against a fake transport
- publish the integer bound in the OpenAPI contract and mirror it in the
  web app's URL parser
- page the popularity index in value order and intersect each page with the
  filter, stopping once the requested page is full, instead of listing every
  matching id and its popularity to rank 24 results
- bound the game_time_to_beats index before reading matching games when a
  duration filter applies, counting both sides so the smaller one is scanned
- record both defects, their cause, and the measured timings in the
  milestone 2 plan

+++ apps/api/src/whattoplaynext_api/catalog/models.py
@@ -133,7 +133,9 @@ class BrowseCriteria(BaseModel):
     genre_ids: tuple[GenreId, ...] = ()
     release_from: date | None = None
     release_to: date | None = None
-    minimum_rating: float | None = Field(default=None, ge=0, le=100)
+    # A whole-number threshold: IGDB's query language rejects a decimal
+    # literal in a where clause, so a fractional minimum could never be served.
+    minimum_rating: int | None = Field(default=None, ge=0, le=100)
     game_mode_ids: tuple[GameModeId, ...] = ()
     duration_kind: DurationKind = DurationKind.NORMAL
     minimum_duration_seconds: int | None = Field(default=None, ge=3600, le=3600000)
